Add padding to align 4096 bytes for gl_strided_buffer

+++ MHD/C_src/KEMO_GL/vartex_array_object_gl.h
@@ -16,7 +16,7 @@
 #define NSIZE_GL_BUFFER  32768
 
 struct gl_strided_buffer{
-	int nsize_buf;
+	long nsize_buf;
 	int istride;
 	
 	int ncomp_buf;
